feat: fix build.gradle and create Application

+++ src/main/resources/db.changelog/v0.0/ddl/2024-01-10--01-create-houses-table.sql
@@ -7,7 +7,7 @@ CREATE TABLE IF NOT EXISTS  houses (
     country VARCHAR(50) NOT NULL,
     city VARCHAR(50) NOT NULL,
     street VARCHAR(50) NOT NULL,
-    number BIGINT NOT NULL,
+    number INTEGER NOT NULL,
     create_date TIMESTAMP NOT NULL,
     UNIQUE (area, country, city, street, number, create_date)
 );
